+++ cmodules/sched_affinity_win32_wrapper.h
@@ -4,9 +4,9 @@
 typedef __int64 cpu_set_t;
 typedef HANDLE pid_t;
 
-static inline int CPU_ISSET(int cpu, cpu_set_t *set)
+static inline int CPU_ISSET(__int64 cpu, cpu_set_t *set)
 {
-	return((*set & (1 << cpu)) ? 1 : 0);
+	return((*set & ((__int64) 1 << cpu)) ? 1 : 0);
 }
 
 static inline int sched_setaffinity(pid_t pid, unsigned int cpusetsize, cpu_set_t *mask)
@@ -17,14 +17,14 @@ static inline int sched_getaffinity(pid_t pid, unsigned int cpusetsize, cpu_set_
 {
 	return(0);
 }
-static inline void CPU_CLR(int cpu, cpu_set_t *set)
+static inline void CPU_CLR(__int64 cpu, cpu_set_t *set)
 {
-	*set &= (~(1 << cpu));
+	*set &= (~((__int64) 1 << cpu));
 }
 
-static inline void CPU_SET(int cpu, cpu_set_t *set)
+static inline void CPU_SET(__int64 cpu, cpu_set_t *set)
 {
-	*set |= (1 << cpu);
+	*set |= ((__int64) 1 << cpu);
 }
 
 static inline void CPU_ZERO(cpu_set_t *set)
